refactor: CompletePage -> MessageLayout 이름 변경
